Fix Hinting Stub Glitch and Enhance
The argparse interface had a bug when started without "run"
Parameter.

Improved the formatting to be more close to PEP 8.

Reworked parser to include "--skip" to skip existing output files
and to ignore non-existing modules. This is to simplify file
handling during automatic creation during a build.

The automatic installer was extracted as an extra patch.
This one should not be subject of a discussion.

Change-Id: I3b6242b7cc459a7080bd8d48f4a37d4582eafb51
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>