Move old glue code to snippets files
Most of the old glue code was directly injected into
the typesystem, so it was possible to add them
as snippets.

There are still a couple of header files that will remain
there, because the include tag does not have the file/snippet
tags.

A few lines of code were modified in favor of "modern" C++,
and good practices.

Task-number: PYSIDE-834
Change-Id: I3072298b16d7280550c6a7f6abae045250663ba6
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>