Fix locking in generate_pyi.py
The file locking was written for Windows in mind.
But file operations are quite different on Linux.

This was no problem during normal builds, but showed
up das a possible racing condition when using the
"--reuse-build" flag.

This version uses a directory to create a lock and has
no platform specific code.

Task-number: PYSIDE-735
Change-Id: I9f27839b0697b49b4dbfea26d6f6949ec466c9d5
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>