Add debug_windows.py script
This script can be used to find out why PySide2 modules
fail to load with various DLL errors like:
- Missing DLL
- Missing symbol in DLL

You can think of it as a Windows version of ldd / LD_DEBUG.

Underneath it uses the cdb.exe command line debugger, and the
gflags.exe tool, both installed with the latest Windows Kit.

The aim is to ask users to run this script when they encounter
PySide2 imports not working on Windows.
The user should then provide the generated log file.

Incindentally it can also be used for any Windows executables,
not just Python.

Consider adapting it to work on Linux and macOS as well
(using LD_DEBUG and something similar on macOS).

Change-Id: Ic7ef37ffd0565495e0774e4db3fc509090607af2
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>