Fix a rare type hint racing condition
When building type hints during the PySide build, it can happen that
the QtMultimedia module already exists, but the QtMultimediaWidgets
module is not yet built.

Since during the build also directories exist alongside with the not yet
existing modules, it can happen that such a directory is picked up by
Python 3, which supports namespace modules without __init__.py .

This case was already handled by the mapping modules, but
QtMultimediaWidgets was directly imported and not checked.

Now the check code has been extracted from the mapping reloader,
and there is no more unchecked module left.

Task-number: PYSIDE-735
Change-Id: I1a1f53525417651005d0759e417082fe71b02773
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>