[reg] Remove createIndex method that uses PyObject*
Even though we have support for Void*, having to import
it as an additional class to be able to call a separate
function, is adding too much complexity so something
that should be easy to do, like creating an index
for an itemmodel.

The regression is a change that replaced the way of doing:
    QAbstractItemModel.createIndex(0, 0, PyObject*)
to
    QAbstractItemModel.createIndex(0, 0, VoidPtr(PyObject*))

which is not really intuitive, and is generating other issues.

Included modification in the registries too.

Change-Id: Ie6112c6baeb4fc3b22fc78e7edeb66aa4a17c22b
Fixes: PYSIDE-883
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>