Improve check when a QFlag<Enum> is found
For the case of the method `addAxis` in the QPolarChart class,
the generated code evaluates first the inherited method:

    QChart::addAxis(QtCharts::QAbstractAxis*,QFlags<Qt::AlignmentFlag>)

instead of the same class method:

    QPolarChart::addAxis(QtCharts::QAbstractAxis*,QtCharts::QPolarChart::PolarOrientation)

The condition to check if an argument was a `QFlags<Qt::AlignmentFlag>`,
was just `PyNumber_Check(...)`, which was too broad, and then any
Enum would pass that condition.

The change was to include an additional condition to check the
type of the EnumType that was passed as argument.

Change-Id: I2eebb6f05c097d170a6d61633698444f03f20b02
Fixes: PYSIDE-898
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>