Fix wrong METADATA for generated wheels
For the 5.12.0 release, we noticed that a couple of strings
in the wheel METADATA were not properly formed.

The first case was the type of the description content:
    'text/markdown',
got interpreted as a tuple  ('text/markdown',),
which was not recognize as a valid format.

The second issue was the summary. It was a two-line string:
    ('Python bindings ... ',
        '....')
which got wrongly parsed, and we ended up missing some title
information, and even getting the parenthesis on the description.

Change-Id: I2f3e8b23e3d9a39355e6eb30ab79f581bde33d30
Fixes: PYSIDE-874
Task-number: QTQAINFRA-2455
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>