Run generate_pyi After Generation of All Binaries
In COIN, there is a racing condition under Windows:

Python opens as many modules as it can, while the build process
is not yet done. This can lead to the situation that a module is
loaded by Python before the Windows Manifest Tool has been
run, and that creates a crash.

We therefore wait when COIN is run, until all binaries have
been created. That means that we are the last process, and
the tool must have been already run.

In non-COIN builds it is more convenient when in case of errors
the generator crashes early.

Task-number: PYSIDE-735
Change-Id: I060dbd54432778f14f74830596f28b4db83a0692
Reviewed-by: Christian Tismer <tismer@stackless.com>