PySide2: Remove non-existent or moved Qt3D classes
Fix warnings:
(3dcore) type 'Qt3DCore::QAbstractEngine' is specified in typesystem, but not defined. This could potentially lead to compilation errors.
(3dcore) type 'Qt3DCore::Quick::QQmlAspectEngine' is specified in typesystem, but not defined. This could potentially lead to compilation errors.
(3drender) type 'Qt3DRender::AssimpImporter' is specified in typesystem, but not defined. This could potentially lead to compilation errors.
(3drender) type 'Qt3DRender::GLTFExporter' is specified in typesystem, but not defined. This could potentially lead to compilation errors.
(3drender) type 'Qt3DRender::GLTFImporter' is specified in typesystem, but not defined. This could potentially lead to compilation errors.

Change-Id: Iee873d69214e8195bfbd72bb664b691dd01d6877
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>