Add script to generate changelogs
This tool generate the changelog for the Qt for Python project.
Since the structure is different from other Qt project,
this script split the changes between the ones to shiboken,
and pyside.

A shiboken change must contain the word "shiboken" in the title,
everything else will be consider a "pyside" change.

The only reported commits are the one associated with a certain
task: Fixes or Task-number.

The option accepts tags, branches or SHAs.

Usage:
    python create_changelog.py -v v5.12.0..v5.12.1 -r 5.12.1
    python create_changelog.py -v v5.11.4..5.13.0 -r 5.13.0 -t minor
    python create_changelog.py -v 9e13465a..5.12 -r 5.12
    python create_changelog.py -v 5.17..6.0 -r 6.0 -t major

The availables  arguments are the following:

  -d DIRECTORY, --directory DIRECTORY
                        Repository directory, the current one is default.
  -v VERSIONS, --versions VERSIONS
                        Tags, branches, or SHA to compare
                        e.g.: v5.12.1..5.12
                              v5.12.0..v5.12.1
                              cebc32a5..5.12
  -r RELEASE, --release RELEASE
                        Release version: e.g.: 5.12.4
  -t TYPE, --type TYPE  Release type: bug-fix (default), minor, or major

There was an issue related to commits that were found twice,
having no task number on the second time.

Added type annotations for the functions.

Sorted the changelog by task-number.

Change-Id: Ia283864fa8add03e136afd75fe7a139ce7e190d8
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>