Make setup.py more robust when invoked in a weird way
setup.py sub-invocations use the command line arguments from the
original setup.py invocation. This means that if setup.py was not
invoked directly, but instead using the "python -c 'code'" approach,
sub-invocations would break because argv[0] would be "-c" instead
of "setup.py".

This was previously seen when building PySide2 in homebrew, and
allegedly when doing an sdist.

Fix setup.py to pass the correct script file name, even when invoked
in such a weird manner.

Change-Id: Ic834d7398098095f74972661b2abc4f4e4c3c8d9
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Alberto Sottile <alby128@gmail.com>