Allow Subclassing PyCFunction in Extension Modules
The signature module tries to figure out if it has to act upon
functions by examining their type.

It was too specific to check for PyCFunction, directly. A user
had built a subclass of PyCFunction and used it with 'inspect.signature',
which then revealed wrong behavior.

This patch removes that restriction.

Change-Id: I7e126ce5750ec5c308cbd1bd1bc4ca4d5eb51e17
Fixes: PYSIDE-950
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>