Prevent Python 3.5 From Crashing The Build
Python 3.5 has a bug that crashes the build.
See the description in the issue tracker.

The cure is to use a more recent contextlib.py and to avoid
a PySide cleanup function that creates the crash.

The problem is not solved for Python 3.5, and it is not clear
if the testbinding module has a hidden bug, too.
But this fix seems to be good enough for the moment.

We should decide if we are going to fix Python 3.5 or abandon
it altogether.

Change-Id: Iacf2237de1f34d2b3cd1d68f1fb5833bdca3fdc2
Fixes: PYSIDE-953
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>