shiboken: Fix the formatting of function signatures
Replace the note directive previously used for \since by the standard
versionadd directive which takes less space.

Move it below the function parameters and indent it along with the
parameters correctly as otherwise the formatting is messed up.

Change writeFunctionSignature() to return a QString which is more
convenient for indenting.

Remove the bool writeDoc parameter from
QtDocGenerator::writeFunction() and spell out the formatting in
writeConstructors() since it all needs to be indented properly (using
the new indenter helper).

Change-Id: I2294ba457ed05f431be295f9d42489aeb6805a8a
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>