Automatically Test Small Example With PyInstaller
A simple hello.py script was modified for running in
PyInstaller and stopping to execute after 2 seconds.

The reason is to test that PyInstaller works correctly
together with the embedded mode of the signature
extension on all platforms.

The script did first not work on Windows. This is now
solved, after an import in pyside2_config.py is fixed.

Currently, there are several configuration errors in COIN.
Errors are therefore skipped in the PyInstaller build.
The test tests only if the generated script works.

Change-Id: I7a1b1e738d523b83cc3fe5beafa7e2579f9c7f48
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>