Improve the Selftest in the Registry
The registry had a duplication of code in the self-test that
led to an oversight.

To prevent that in the future, the error handling has been
rewritten, and an extended self-test was written that checks
the correctness of every case.

A truth table has beed added as a reference for questions.

Change-Id: I4570b2388c9b179b33c9520a484e09b9ccb68918
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>