Shiboken: Add "noexcept" to wrapper when wrapped function is noexcept
The clangparser was already recording the exception specifier for
wrapped functions, but the code generator did not transfer the type
specifier to the wrapper function. With modern compilers this leads
to compile errors as the exception specifier is now part of the
type. Noexcept is now added whenever the clangparser determines the
source function is noexcept.

Change-Id: I1f68981d997fcf4f007569c6d682e3dcf68eca60
Fixes: PYSIDE-977
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>