Amend the Embedding Patch for cx_Freeze
cx_Freeze has the bug that it copies the "files.dir" folder
of shiboken, but does not insert the Python files.

As a quick fix, instead of testing only for existence of this
directory, we now check for existence of the loader.py file.
It can be assumed that all other Python files will be there
as well.

We could go even further and test-load all files. But then
it becomes questionable if we should better always embed.

Change-Id: Ib9553941c6a658fb20cb85d22f78431f99d88734
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit b84b5a2f35bea95545c19ee313be5ec5db19eca6)
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>