Doc: Add descriptions, pages, and moving modules
Having all the Qt modules on the main page is too verbose.
I created four sections with a couple of them and added
more context to clarify when to use them.

All the modules are still listed in a new page called `modules`.

I included an `overview` page with some general information from
our wiki: compatibility among platforms, an example and the FAQ.

Task-number: PYSIDE-691
Change-Id: Ic9e2519aa2caa9c8a645633c2c1772e4844b5ec5
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>