Add contextlib to Preloaded Modules for Embedding with Python 2
While developing the embedding for installer support, some functionality
was moved from loader.py to signature_bootstrap.py, especially a "with"
statement, that needed a contextlib import.

When using PyInstaller or cxFreeze, this gave a problem when running
in virtualenv and Python 2. We extended the pre-loading to contextlib.

It is not completely clear why Python 2 needs that and Python 3 doesn't.

The problem did not show up without virtualenv. Perhaps some different
installed packages have hidden that problem, because they loaded the
contextlib, themselves. This does not happen with a freshly installed
virtual environment.

Change-Id: I2188cb67129e596f6492d0f5dc9b303a67bfe870
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>