Use correct environment variable to detect when running in Coin
The code for shipping MSVC redist dlls used to check for the
"QTEST_ENVIRONMENT" environment variable to decide when to
download and extract the libraries. This was incorrect, because
it is only set in a Coin test phase, and not when building
the project.

Revert to using the old "COIN_LAUNCH_PARAMETERS" environment variable
which is set during the Coin build phase.

Amends 9e13465a992a5079c0a038e5f9417896dcfd6f3a .

Fixes: PYSIDE-993
Change-Id: I96d1e010ffc4b3a5951bdd28699f657e3a5f4a3f
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>