shiboken: Fix code injection not working for operator functions
Searching for the function modifications of the operator function by
signature failed due to the internally changed signature.

Store the function modification list of user-added-functions directly
in struct AddedFunction instead of the type entry of the containing
class.

In AbstractMetaFunction, replace the bool m_userAdded flag by a shared
pointer to the AddedFunction and use that to retrieve the
modifications (injected code snippets) for the user-added functions
instead of searching for them by function signature.

Task-number: PYSIDE-995
Change-Id: Ic4d0b257f141a450df26563d33beb397b6209d91
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>