Provide a spec file for hello.app used by wheel_tester
The PyInstaller test was done with minimalistic options to keep
things as simple as possible.

This had the side effect that on certain Python versions and
platforms things were touched that created errors, although
those things were nowhere in use.

It turned out that the generated hello_app.spec file contains
the result of an in-depth analysis which touches so many
non-existing modules. When using just the generated spec file and
no other option, the analysis is almost skipped and the generation
of an executable avoids further problems and is faster.

This patch removes all arguments and uses a spec file instead.
The arguments are kept in a comment, just in case.

Change-Id: I17c43a588bc6f95f62fc3104a047dc08e8d8e70d
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>