Fix pyi generation on Windows
When trying to generate the pyi files on Windows, the python modules
failed to be imported due to Qt dlls not being in PATH. This happened
because we overrode PATH instead of prepending to it.

When generating pyi files on Windows, make sure to prepend new values
to PATH while also keeping the old values.

Amends 97df448edbc035a2f531508a5cfe5739fb2de18d

Change-Id: I74c374c3a9687169dd4af0ed232bda76a0267fc4
Reviewed-by: Christian Tismer <tismer@stackless.com>