Support the qApp macro in "scriptable application"
Renamed from "Fix scriptable application to support the qApp macro"
because qApp was improved instead of scriptable application.

The qApp macro needed some extra effort to support the
qApp "macro" which is only defined in the Python wrappers.

I took some generated code, created a QApplication instance
in Python and used then reduced generated code to get at the
object and adjust the refcount.

This solution was then rejected, because I can do better,
and in fact, scriptable application now has a correct qApp
macro too, without any change to scriptable application.

The central idea was to look into the module init function
at import time and to see if a Q*Application already exists.
I was not aware of that import. Many thanks for the rejection! :-)

Update..
--------
After many attempts to make the qApp variable correctly behave
like always, I recognized that pre-existing Q*Application instances
have no wrappers or constructors at all! With that, it is not
possible to create a sophisticated qApp macro as a singleton
variable in the desired way.

Fortunately, this is also not necessary, because a C++ Q*Application
cannot be deleted from Python, and there is no point in supporting
more that a simple variable. So in case of a pre-existing instance,
the qApp variable now gets redirected to that instance.

A small test was added to application_test.py that is triggered by
an import.

A weird effect when "qApp" was typed interactively before calling
"QApplication()" was fixed, too.

Change-Id: Ic69dd6a21c964838a90f63e316d299b62a54d612
Fixes: PYSIDE-571
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>