Add optional support for types in QSettings::value
This add an optional named parameter to the function value()
to automatically cast the type that is being returned
by the function.

An example of this situation could be an ini file that contains
the value of a one-element list:

    settings.setValue('var', ['a'])

The the ini file will be:

    [General]
    var=a # we cannot know that this is a list!

Once we read it, we could specify if we want
the default behavior, a str, or to cast the output
to a list.

    settings.value('var') # Will get "a"
    settings.value('var', type=list) # Will get ["a"]

The cppgenerator was modified to add a verification step
before trying to get the named parameter, since it could
be optional and having one named parameter was assumming
that all of them were provided.

Change-Id: I8f379debea86b42cf89019d432e990084c9e6614
Fixes: PYSIDE-1010
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Tismer <tismer@stackless.com>